feat: added functionality for user management within an organisation (WIP) (#31)

+++ pages/api/frontend/v0.1/orgs/[orgId]/users.tsx
@@ -50,7 +50,7 @@ export default async function handler(
         }
     });
 
-    if (userInOrg?.role !== "ADMIN") {
+    if (userInOrg?.role !== "ADMIN" && userInOrg?.role !== "USER") {
         // if user has no business with this organisation, return a 404
         res.status(404).json({ message: 'no organisation found with id ' + req.query.orgId });
         return;
@@ -86,6 +86,10 @@ export default async function handler(
 
         case 'DELETE':
             try {
+                if (userInOrg?.role === "USER") {
+                    res.status(403).json({ message: 'you are not allowed to delete user with id ' + req.query.userId + ' from organisation with id ' + req.query.orgId });
+                    return;
+                }
                 const deletedUserInOrg = await prisma.usersInOrganisations.delete({
                     where: {
                         orgId_userId: {
@@ -103,27 +107,73 @@ export default async function handler(
             }
             break;
 
-        case 'PUT':
-            try {
-                const updatedApp = await prisma.usersInOrganisations.update({
-                    where: {
-                        orgId_userId: {
-                            userId: Number(req.query.userId),
-                            orgId: Number(req.query.orgId),
-                        }
-                    },
-                    data: {
-                        role: req.body.role
+            case 'PUT':
+                try {
+                    if (userInOrg?.role === "USER") {
+                        res.status(403).json({ message: 'you are not allowed to update user with id ' + req.query.userId + ' from organisation with id ' + req.query.orgId });
+                        return;
                     }
-                });
-
-                res.status(201).json(updatedApp)
-            } catch(e) {
-                if (e instanceof Prisma.PrismaClientKnownRequestError) {
-                    res.status(404).json({ message: 'no user with id ' + req.query.userId + ' found in organisation with id ' + req.query.appId });
+                    const updatedApp = await prisma.usersInOrganisations.update({
+                        where: {
+                            orgId_userId: {
+                                userId: Number(req.query.userId),
+                                orgId: Number(req.query.orgId),
+                            }
+                        },
+                        data: {
+                            role: req.body.role
+                        }
+                    });
+    
+                    res.status(201).json(updatedApp)
+                } catch(e) {
+                    if (e instanceof Prisma.PrismaClientKnownRequestError) {
+                        res.status(404).json({ message: 'no user with id ' + req.query.userId + ' found in organisation with id ' + req.query.appId });
+                    }
                 }
-            }
-            break;
+                break;
+
+            case 'POST':
+                try {
+                    if (userInOrg?.role === "USER") {
+                        res.status(403).json({ message: 'you are not allowed to add user with email ' + req.body.email + ' to organisation with id ' + req.query.orgId });
+                        return;
+                    }
+
+                    const user = await prisma.user.findFirst({
+                        where: {
+                            email: req.body.email
+                        }
+                    });
+
+                    if (user && user.id) {
+                        const org = await prisma.usersInOrganisations.create({
+                            data: {
+                                user: {
+                                    connect: {
+                                        id: user.id
+                                    }
+                                },
+                                role: 'USER',
+                                org: {
+                                    create: {
+                                        name: req.body.name
+                                    }
+                                }
+                            }
+                        })
+                        res.status(201).json(org);
+                        return;
+                    }
+                    res.status(400).json({ message: 'no user found for email' });
+                    return;
+                    
+                } catch(e) {
+                    if (e instanceof Prisma.PrismaClientKnownRequestError) {
+                        res.status(404).json({ message: 'no user with id ' + req.query.userId + ' found in organisation with id ' + req.query.appId });
+                    }
+                }
+                break;
 
         default:
             res.status(405).json({ message: 'method not allowed' });
